+++ src/main/kotlin/com/speedyteller/reporting/api/domain/model/FXTransaction.kt
@@ -3,7 +3,8 @@ package com.speedyteller.reporting.api.domain.model
 import java.math.BigDecimal
 
 data class FXTransaction(
-    val id: Int? = null,
+
+    val id: Long? = null,
     val originalAmount : BigDecimal? = null,
     val originalCurrency : String? = null
 )
