+++ src/main/kotlin/com/speedyteller/reporting/api/domain/model/InstantPaymentNotification.kt
@@ -1,7 +1,8 @@
 package com.speedyteller.reporting.api.domain.model
 
 data class InstantPaymentNotification(
-    val id: Int? = null,
+
+    val id: Long? = null,
     val transactionId: Int? = null,
     val received: Boolean? = null
 )
