+++ src/main/kotlin/com/speedyteller/reporting/api/domain/model/Merchant.kt
@@ -1,6 +1,7 @@
 package com.speedyteller.reporting.api.domain.model
 
 data class Merchant(
-    val id: Int? = null,
+
+    val id: Long? = null,
     val name: String? = null
 )
